+++ flowcraft/generator/inspect.py
@@ -1540,7 +1540,8 @@ class NextflowInspector:
         # Get hash from the current working dir and hostname
         workdir = self.workdir.encode("utf8")
         hostname = socket.gethostname().encode("utf8")
-        dir_hash = hashlib.md5(workdir + hostname)
+        timestamp = str(time.time()).encode("utf8")
+        dir_hash = hashlib.md5(workdir + hostname + timestamp)
 
         return pipeline_hash.hexdigest() + dir_hash.hexdigest()